chore: removing dead (exploratory) code
I had tried using a macro to pull out all of the struct field names into
an array.  To assist in automation, but the macro can't gaurantee order
nor, am I aware, of a way to directly associatte the struct fields with
the name: opening up the possibility of misalignment.

(Note: I believe this was of use because we were also referencing
template values of the same name, as strings, and this would have helped
group the two together.  Future: a macro that pulls the template values
out directly would be great and it could populate both the struct field
and the string:struct association.)